Add files via upload
"Electron App with C++ Integration

This project is an example of an Electron app that uses C++ code to interact with the operating system or other native libraries. The app uses the node-addon-api and ffi-napi libraries to interface with the C++ code from JavaScript.

To build and run the app, simply clone the repository and run npm install followed by npm start. The app will be built and launched automatically.

Note: This project requires a C++ compiler to be installed on your system in order to build the native add-on."

@amultomar
insta :- @rudra_.tomar